Contabilidad: importes en céntimos enteros
El debe y el haber pasan de decimal(12,2) a bigint de céntimos, y el cuadre se
comprueba con una igualdad exacta en vez de comparar redondeos.

Las columnas van con signo a propósito: siendo UNSIGNED, MariaDB revienta en
cualquier resta a nivel de fila (debe - haber, SUM(debe - haber), el saldo
acumulado del mayor). El no-negativo se valida en la aplicación.

Los céntimos se quedan así en todo el modelo y solo se dividen al presentar: los
agregados de SQL (withSum) se saltan los casts de Eloquent, y si el atributo
devolviera euros, el mismo nombre tendría dos unidades según por dónde se leyera.

+++ database/migrations/2026_08_04_070000_create_apunte_contables_table.php
@@ -15,8 +15,12 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('asiento_contable_id')->index('apunte_contables_asiento_contable_id_foreign');
             $table->unsignedBigInteger('cuenta_contable_id')->index('apunte_contables_cuenta_contable_id_foreign');
-            $table->decimal('debe', 12, 2)->default(0);
-            $table->decimal('haber', 12, 2)->default(0);
+            // Importes en céntimos, enteros. Con signo a propósito: con columnas UNSIGNED,
+            // MariaDB revienta en cualquier resta a nivel de fila — SUM(debe - haber),
+            // debe - haber, o el saldo acumulado del mayor con una window function.
+            // El no-negativo se valida en la aplicación.
+            $table->bigInteger('debe')->default(0);
+            $table->bigInteger('haber')->default(0);
             $table->string('concepto', 255)->nullable();
             $table->timestamps();
